FateTools/requirements.txt
asgiref==3.5.2
cffi==1.15.0
cryptography==37.0.2
Django==4.0.4
pycparser==2.21
python-gnupg==0.4.9
sqlparse==0.4.2
